+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+2.4.0
+
+  Match X route behavior
+  Screenshot with Chromium
+  Use forwarded client IP
+
 2.3.0
 
   Migrate skateboard 5.6